push

+++ application/models/M_pelajaran.php
@@ -12,9 +12,16 @@ class M_pelajaran extends CI_Model {
 
 	public function insert()
 	{
-		return $this->db->insert('pelajaran', [
+		$this->db->insert('pelajaran', [
 			'nama_pelajaran' => $this->input->post('nama_pelajaran')
 		]);
+		$id = $this->db->insert_id();
+		foreach ($this->M_siswa->get()->result() as $key) {
+			if ($this->M_nilai->get(['nilai.id_pelajaran' => $id, 'nilai.id_siswa' => $key->id_siswa])->num_rows() == 0) {
+				$this->db->insert('nilai', ['id_siswa' => $key->id_siswa, 'id_pelajaran' => $id]);
+			}
+		}
+		return 1;
 	}
 
 	public function update($id)
@@ -26,7 +33,11 @@ class M_pelajaran extends CI_Model {
 
 	public function delete($id)
 	{
-		return $this->db->delete('pelajaran', ['md5(id_pelajaran)' => $id]);
+		if ($this->M_nilai->get(['md5(nilai.id_pelajaran)' => $id])->num_rows() > 0) {
+			return false;
+		}else{
+			return $this->db->delete('pelajaran', ['md5(id_pelajaran)' => $id]);
+		}
 	}
 
 }
